Tests/LibWeb/Text/expected/DOM/Element-aria-element-reflection-cache.txt

first === second: true
length: 1
element id: target
